+++ example/android/app/src/main/kotlin/com/example/flutter_pty_new_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_pty_example
+package com.example.flutter_pty_new_example
 
 import io.flutter.embedding.android.FlutterActivity